+++ scheduler/scheduler.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	. "project/database"
+	. "project/tables"
 	"time"
 
 	q "github.com/daviddengcn/go-villa"
